backend/requirements.txt
fastapi
uvicorn
gunicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-psycopg2
